feat(con-384): migrate link button group to typescript
