+++ src/components/LinkButtonGroup/LinkButtonGroup.spec.tsx
@@ -1,0 +1,9 @@
+// import React from 'react';
+// import { render, screen } from '@testing-library/react';
+// import { LinkButtonGroup } from './LinkButtonGroup';
+// import { Hyperlink } from '~/components/Hyperlink';
+// import { Button } from '~/components/Button';
+
+describe('<LinkButtonGroup />', () => {
+  it.todo('figure out out an efficient way to test this');
+});
